+++ dd-java-agent/appsec/src/main/java/com/datadog/appsec/gateway/GatewayBridge.java
@@ -1,49 +1,91 @@
 package com.datadog.appsec.gateway;
 
+import static com.datadog.appsec.event.data.MapDataBundle.Builder.CAPACITY_0_2;
+import static com.datadog.appsec.event.data.MapDataBundle.Builder.CAPACITY_6_10;
+import static com.datadog.appsec.gateway.AppSecRequestContext.DEFAULT_REQUEST_HEADERS_ALLOW_LIST;
+import static com.datadog.appsec.gateway.AppSecRequestContext.REQUEST_HEADERS_ALLOW_LIST;
+import static com.datadog.appsec.gateway.AppSecRequestContext.RESPONSE_HEADERS_ALLOW_LIST;
+
+import com.datadog.appsec.AppSecSystem;
 import com.datadog.appsec.api.security.ApiSecurityRequestSampler;
 import com.datadog.appsec.config.TraceSegmentPostProcessor;
 import com.datadog.appsec.event.EventProducerService;
+import com.datadog.appsec.event.EventProducerService.DataSubscriberInfo;
+import com.datadog.appsec.event.ExpiredSubscriberInfoException;
 import com.datadog.appsec.event.data.Address;
+import com.datadog.appsec.event.data.DataBundle;
 import com.datadog.appsec.event.data.KnownAddresses;
-import com.datadog.appsec.gateway.callbacks.DatabaseConnectionCallback;
-import com.datadog.appsec.gateway.callbacks.DatabaseSqlQueryCallback;
-import com.datadog.appsec.gateway.callbacks.GraphqlServerRequestMessageCallback;
-import com.datadog.appsec.gateway.callbacks.GrpcServerMethodCallback;
-import com.datadog.appsec.gateway.callbacks.GrpcServerRequestMessageCallback;
-import com.datadog.appsec.gateway.callbacks.MethodAndRawURICallback;
-import com.datadog.appsec.gateway.callbacks.NewRequestHeaderCallback;
-import com.datadog.appsec.gateway.callbacks.RequestBodyDoneCallback;
-import com.datadog.appsec.gateway.callbacks.RequestBodyProcessedCallback;
-import com.datadog.appsec.gateway.callbacks.RequestBodyStartCallback;
-import com.datadog.appsec.gateway.callbacks.RequestClientSocketAddressCallback;
-import com.datadog.appsec.gateway.callbacks.RequestEndedCallBack;
-import com.datadog.appsec.gateway.callbacks.RequestHeadersDoneCallback;
-import com.datadog.appsec.gateway.callbacks.RequestInferredClientAddressCallback;
-import com.datadog.appsec.gateway.callbacks.RequestPathParamsCallback;
-import com.datadog.appsec.gateway.callbacks.RequestStartedCallback;
-import com.datadog.appsec.gateway.callbacks.ResponseHeaderCallback;
-import com.datadog.appsec.gateway.callbacks.ResponseHeaderDoneCallback;
-import com.datadog.appsec.gateway.callbacks.ResponseStartedCallback;
+import com.datadog.appsec.event.data.MapDataBundle;
+import com.datadog.appsec.event.data.ObjectIntrospection;
+import com.datadog.appsec.event.data.SingletonDataBundle;
+import com.datadog.appsec.report.AppSecEvent;
+import com.datadog.appsec.report.AppSecEventWrapper;
+import com.datadog.appsec.stack_trace.StackTraceCollection;
+import com.datadog.appsec.util.ObjectFlattener;
+import datadog.trace.api.Config;
+import datadog.trace.api.function.TriConsumer;
+import datadog.trace.api.function.TriFunction;
 import datadog.trace.api.gateway.Events;
+import datadog.trace.api.gateway.Flow;
+import datadog.trace.api.gateway.IGSpanInfo;
+import datadog.trace.api.gateway.RequestContext;
+import datadog.trace.api.gateway.RequestContextSlot;
 import datadog.trace.api.gateway.SubscriptionService;
+import datadog.trace.api.http.StoredBodySupplier;
+import datadog.trace.api.internal.TraceSegment;
+import datadog.trace.api.telemetry.RuleType;
+import datadog.trace.api.telemetry.WafMetricCollector;
+import datadog.trace.bootstrap.instrumentation.api.Tags;
+import datadog.trace.bootstrap.instrumentation.api.URIDataAdapter;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.function.Function;
+import java.util.regex.Pattern;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /** Bridges the instrumentation gateway and the reactive engine. */
 public class GatewayBridge {
   private static final Events<AppSecRequestContext> EVENTS = Events.get();
+
+  private static final Logger log = LoggerFactory.getLogger(GatewayBridge.class);
+
+  private static final Pattern QUERY_PARAM_VALUE_SPLITTER = Pattern.compile("=");
+  private static final Pattern QUERY_PARAM_SPLITTER = Pattern.compile("&");
+  private static final Map<String, List<String>> EMPTY_QUERY_PARAMS = Collections.emptyMap();
+
+  /** User tracking tags that will force the collection of request headers */
+  private static final String[] USER_TRACKING_TAGS = {
+    "appsec.events.users.login.success.track", "appsec.events.users.login.failure.track"
+  };
 
   private final SubscriptionService subscriptionService;
   private final EventProducerService producerService;
   private final ApiSecurityRequestSampler requestSampler;
   private final List<TraceSegmentPostProcessor> traceSegmentPostProcessors;
 
-  private final SubscribersCache subscribersCache;
+  // subscriber cache
+  private volatile DataSubscriberInfo initialReqDataSubInfo;
+  private volatile DataSubscriberInfo rawRequestBodySubInfo;
+  private volatile DataSubscriberInfo requestBodySubInfo;
+  private volatile DataSubscriberInfo pathParamsSubInfo;
+  private volatile DataSubscriberInfo respDataSubInfo;
+  private volatile DataSubscriberInfo grpcServerMethodSubInfo;
+  private volatile DataSubscriberInfo grpcServerRequestMsgSubInfo;
+  private volatile DataSubscriberInfo graphqlServerRequestMsgSubInfo;
+  private volatile DataSubscriberInfo requestEndSubInfo;
+  private volatile DataSubscriberInfo dbSqlQuerySubInfo;
 
   public GatewayBridge(
       SubscriptionService subscriptionService,
@@ -54,7 +96,6 @@ public class GatewayBridge {
     this.producerService = producerService;
     this.requestSampler = requestSampler;
     this.traceSegmentPostProcessors = traceSegmentPostProcessors;
-    this.subscribersCache = new SubscribersCache();
   }
 
   public void init() {
@@ -63,75 +104,726 @@ public class GatewayBridge {
         IGAppSecEventDependencies.additionalIGEventTypes(
             producerService.allSubscribedDataAddresses());
 
-    subscriptionService.registerCallback(events.requestStarted(), new RequestStartedCallback());
+    subscriptionService.registerCallback(
+        events.requestStarted(),
+        () -> {
+          if (!AppSecSystem.isActive()) {
+            return RequestContextSupplier.EMPTY;
+          }
+          return new RequestContextSupplier();
+        });
 
     subscriptionService.registerCallback(
         events.requestEnded(),
-        new RequestEndedCallBack(
-            requestSampler, subscribersCache, producerService, traceSegmentPostProcessors));
+        (RequestContext ctx_, IGSpanInfo spanInfo) -> {
+          AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+          if (ctx == null) {
+            return NoopFlow.INSTANCE;
+          }
+
+          maybeExtractSchemas(ctx);
+
+          // WAF call
+          ctx.closeAdditive();
+
+          TraceSegment traceSeg = ctx_.getTraceSegment();
+
+          // AppSec report metric and events for web span only
+          if (traceSeg != null) {
+            traceSeg.setTagTop("_dd.appsec.enabled", 1);
+            traceSeg.setTagTop("_dd.runtime_family", "jvm");
+
+            Collection<AppSecEvent> collectedEvents = ctx.transferCollectedEvents();
+
+            for (TraceSegmentPostProcessor pp : this.traceSegmentPostProcessors) {
+              pp.processTraceSegment(traceSeg, ctx, collectedEvents);
+            }
+
+            // If detected any events - mark span at appsec.event
+            if (!collectedEvents.isEmpty()) {
+              // Set asm keep in case that root span was not available when events are detected
+              traceSeg.setTagTop(Tags.ASM_KEEP, true);
+              traceSeg.setTagTop(Tags.PROPAGATED_APPSEC, true);
+              traceSeg.setTagTop("appsec.event", true);
+              traceSeg.setTagTop("network.client.ip", ctx.getPeerAddress());
+
+              // Reflect client_ip as actor.ip for backward compatibility
+              Object clientIp = spanInfo.getTags().get(Tags.HTTP_CLIENT_IP);
+              if (clientIp != null) {
+                traceSeg.setTagTop("actor.ip", clientIp);
+              }
+
+              // Report AppSec events via "_dd.appsec.json" tag
+              AppSecEventWrapper wrapper = new AppSecEventWrapper(collectedEvents);
+              traceSeg.setDataTop("appsec", wrapper);
+
+              // Report collected request and response headers based on allow list
+              writeRequestHeaders(traceSeg, REQUEST_HEADERS_ALLOW_LIST, ctx.getRequestHeaders());
+              writeResponseHeaders(traceSeg, RESPONSE_HEADERS_ALLOW_LIST, ctx.getResponseHeaders());
+
+              // Report collected stack traces
+              StackTraceCollection stackTraceCollection = ctx.transferStackTracesCollection();
+              if (stackTraceCollection != null) {
+                Object flatStruct = ObjectFlattener.flatten(stackTraceCollection);
+                if (flatStruct != null) {
+                  traceSeg.setMetaStructTop("_dd.stack", flatStruct);
+                }
+              }
+            } else if (hasUserTrackingEvent(traceSeg)) {
+              // Report all collected request headers on user tracking event
+              writeRequestHeaders(traceSeg, REQUEST_HEADERS_ALLOW_LIST, ctx.getRequestHeaders());
+            } else {
+              // Report minimum set of collected request headers
+              writeRequestHeaders(
+                  traceSeg, DEFAULT_REQUEST_HEADERS_ALLOW_LIST, ctx.getRequestHeaders());
+            }
+            // If extracted any Api Schemas - commit them
+            if (!ctx.commitApiSchemas(traceSeg)) {
+              log.debug("Unable to commit, api security schemas and will be skipped");
+            }
+
+            if (ctx.isBlocked()) {
+              WafMetricCollector.get().wafRequestBlocked();
+            } else if (!collectedEvents.isEmpty()) {
+              WafMetricCollector.get().wafRequestTriggered();
+            } else {
+              WafMetricCollector.get().wafRequest();
+            }
+          }
+
+          ctx.close();
+          return NoopFlow.INSTANCE;
+        });
 
     subscriptionService.registerCallback(EVENTS.requestHeader(), new NewRequestHeaderCallback());
     subscriptionService.registerCallback(
-        EVENTS.requestHeaderDone(),
-        new RequestHeadersDoneCallback(subscribersCache, producerService));
+        EVENTS.requestHeaderDone(), new RequestHeadersDoneCallback());
 
     subscriptionService.registerCallback(
-        EVENTS.requestMethodUriRaw(),
-        new MethodAndRawURICallback(subscribersCache, producerService));
+        EVENTS.requestMethodUriRaw(), new MethodAndRawURICallback());
 
-    subscriptionService.registerCallback(EVENTS.requestBodyStart(), new RequestBodyStartCallback());
+    subscriptionService.registerCallback(
+        EVENTS.requestBodyStart(),
+        (RequestContext ctx_, StoredBodySupplier supplier) -> {
+          AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+          if (ctx == null) {
+            return null;
+          }
+
+          ctx.setStoredRequestBodySupplier(supplier);
+          return null;
+        });
 
     if (additionalIGEvents.contains(EVENTS.requestPathParams())) {
       subscriptionService.registerCallback(
           EVENTS.requestPathParams(),
-          new RequestPathParamsCallback(subscribersCache, producerService));
+          (ctx_, data) -> {
+            AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+            if (ctx == null || ctx.isPathParamsPublished()) {
+              return NoopFlow.INSTANCE;
+            }
+            ctx.setPathParamsPublished(true);
+
+            while (true) {
+              DataSubscriberInfo subInfo = pathParamsSubInfo;
+              if (subInfo == null) {
+                subInfo = producerService.getDataSubscribers(KnownAddresses.REQUEST_PATH_PARAMS);
+                pathParamsSubInfo = subInfo;
+              }
+              if (subInfo == null || subInfo.isEmpty()) {
+                return NoopFlow.INSTANCE;
+              }
+              DataBundle bundle =
+                  new SingletonDataBundle<>(KnownAddresses.REQUEST_PATH_PARAMS, data);
+              try {
+                GatewayContext gwCtx = new GatewayContext(false);
+                return producerService.publishDataEvent(subInfo, ctx, bundle, gwCtx);
+              } catch (ExpiredSubscriberInfoException e) {
+                pathParamsSubInfo = null;
+              }
+            }
+          });
     }
 
     subscriptionService.registerCallback(
-        EVENTS.requestBodyDone(), new RequestBodyDoneCallback(subscribersCache, producerService));
+        EVENTS.requestBodyDone(),
+        (RequestContext ctx_, StoredBodySupplier supplier) -> {
+          AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+          if (ctx == null || ctx.isRawReqBodyPublished()) {
+            return NoopFlow.INSTANCE;
+          }
+          ctx.setRawReqBodyPublished(true);
+
+          while (true) {
+            DataSubscriberInfo subInfo = rawRequestBodySubInfo;
+            if (subInfo == null) {
+              subInfo = producerService.getDataSubscribers(KnownAddresses.REQUEST_BODY_RAW);
+              rawRequestBodySubInfo = subInfo;
+            }
+            if (subInfo == null || subInfo.isEmpty()) {
+              return NoopFlow.INSTANCE;
+            }
+
+            CharSequence bodyContent = supplier.get();
+            if (bodyContent == null || bodyContent.length() == 0) {
+              return NoopFlow.INSTANCE;
+            }
+            DataBundle bundle =
+                new SingletonDataBundle<>(KnownAddresses.REQUEST_BODY_RAW, bodyContent);
+            try {
+              GatewayContext gwCtx = new GatewayContext(false);
+              return producerService.publishDataEvent(subInfo, ctx, bundle, gwCtx);
+            } catch (ExpiredSubscriberInfoException e) {
+              rawRequestBodySubInfo = null;
+            }
+          }
+        });
 
     if (additionalIGEvents.contains(EVENTS.requestBodyProcessed())) {
       subscriptionService.registerCallback(
           EVENTS.requestBodyProcessed(),
-          new RequestBodyProcessedCallback(subscribersCache, producerService));
+          (RequestContext ctx_, Object obj) -> {
+            AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+            if (ctx == null) {
+              return NoopFlow.INSTANCE;
+            }
+
+            if (ctx.isConvertedReqBodyPublished()) {
+              log.debug(
+                  "Request body already published; will ignore new value of type {}",
+                  obj.getClass());
+              return NoopFlow.INSTANCE;
+            }
+            ctx.setConvertedReqBodyPublished(true);
+
+            while (true) {
+              DataSubscriberInfo subInfo = requestBodySubInfo;
+              if (subInfo == null) {
+                subInfo = producerService.getDataSubscribers(KnownAddresses.REQUEST_BODY_OBJECT);
+                requestBodySubInfo = subInfo;
+              }
+              if (subInfo == null || subInfo.isEmpty()) {
+                return NoopFlow.INSTANCE;
+              }
+              DataBundle bundle =
+                  new SingletonDataBundle<>(
+                      KnownAddresses.REQUEST_BODY_OBJECT, ObjectIntrospection.convert(obj));
+              try {
+                GatewayContext gwCtx = new GatewayContext(false);
+                return producerService.publishDataEvent(subInfo, ctx, bundle, gwCtx);
+              } catch (ExpiredSubscriberInfoException e) {
+                requestBodySubInfo = null;
+              }
+            }
+          });
     }
 
     subscriptionService.registerCallback(
         EVENTS.requestClientSocketAddress(),
-        new RequestClientSocketAddressCallback(subscribersCache, producerService));
+        (ctx_, ip, port) -> {
+          AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+          if (ctx == null || ctx.isReqDataPublished()) {
+            return NoopFlow.INSTANCE;
+          }
+          ctx.setPeerAddress(ip);
+          ctx.setPeerPort(port);
+          return maybePublishRequestData(ctx);
+        });
 
     subscriptionService.registerCallback(
-        EVENTS.requestInferredClientAddress(), new RequestInferredClientAddressCallback());
+        EVENTS.requestInferredClientAddress(),
+        (ctx_, ip) -> {
+          AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+          if (ctx != null) {
+            ctx.setInferredClientIp(ip);
+          }
+          return NoopFlow.INSTANCE; // expected to be called before requestClientSocketAddress
+        });
 
     subscriptionService.registerCallback(
-        EVENTS.responseStarted(), new ResponseStartedCallback(subscribersCache, producerService));
+        EVENTS.responseStarted(),
+        (ctx_, status) -> {
+          AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+          if (ctx == null || ctx.isRespDataPublished()) {
+            return NoopFlow.INSTANCE;
+          }
+          ctx.setResponseStatus(status);
+          return maybePublishResponseData(ctx);
+        });
 
-    subscriptionService.registerCallback(EVENTS.responseHeader(), new ResponseHeaderCallback());
-
+    subscriptionService.registerCallback(
+        EVENTS.responseHeader(),
+        (ctx_, name, value) -> {
+          AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+          if (ctx != null) {
+            ctx.addResponseHeader(name, value);
+          }
+        });
     subscriptionService.registerCallback(
         EVENTS.responseHeaderDone(),
-        new ResponseHeaderDoneCallback(subscribersCache, producerService));
+        ctx_ -> {
+          AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+          if (ctx == null || ctx.isRespDataPublished()) {
+            return NoopFlow.INSTANCE;
+          }
+          ctx.finishResponseHeaders();
+          return maybePublishResponseData(ctx);
+        });
 
     subscriptionService.registerCallback(
-        EVENTS.grpcServerMethod(), new GrpcServerMethodCallback(subscribersCache, producerService));
+        EVENTS.grpcServerMethod(),
+        (ctx_, method) -> {
+          AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+          if (ctx == null || method == null || method.isEmpty()) {
+            return NoopFlow.INSTANCE;
+          }
+          while (true) {
+            DataSubscriberInfo subInfo = grpcServerMethodSubInfo;
+            if (subInfo == null) {
+              subInfo = producerService.getDataSubscribers(KnownAddresses.GRPC_SERVER_METHOD);
+              grpcServerMethodSubInfo = subInfo;
+            }
+            if (subInfo == null || subInfo.isEmpty()) {
+              return NoopFlow.INSTANCE;
+            }
+            DataBundle bundle =
+                new SingletonDataBundle<>(KnownAddresses.GRPC_SERVER_METHOD, method);
+            try {
+              GatewayContext gwCtx = new GatewayContext(true);
+              return producerService.publishDataEvent(subInfo, ctx, bundle, gwCtx);
+            } catch (ExpiredSubscriberInfoException e) {
+              grpcServerMethodSubInfo = null;
+            }
+          }
+        });
 
     subscriptionService.registerCallback(
         EVENTS.grpcServerRequestMessage(),
-        new GrpcServerRequestMessageCallback(subscribersCache, producerService));
+        (ctx_, obj) -> {
+          AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+          if (ctx == null) {
+            return NoopFlow.INSTANCE;
+          }
+          while (true) {
+            DataSubscriberInfo subInfo = grpcServerRequestMsgSubInfo;
+            if (subInfo == null) {
+              subInfo =
+                  producerService.getDataSubscribers(KnownAddresses.GRPC_SERVER_REQUEST_MESSAGE);
+              grpcServerRequestMsgSubInfo = subInfo;
+            }
+            if (subInfo == null || subInfo.isEmpty()) {
+              return NoopFlow.INSTANCE;
+            }
+            Object convObj = ObjectIntrospection.convert(obj);
+            DataBundle bundle =
+                new SingletonDataBundle<>(KnownAddresses.GRPC_SERVER_REQUEST_MESSAGE, convObj);
+            try {
+              GatewayContext gwCtx = new GatewayContext(true);
+              return producerService.publishDataEvent(subInfo, ctx, bundle, gwCtx);
+            } catch (ExpiredSubscriberInfoException e) {
+              grpcServerRequestMsgSubInfo = null;
+            }
+          }
+        });
 
     subscriptionService.registerCallback(
         EVENTS.graphqlServerRequestMessage(),
-        new GraphqlServerRequestMessageCallback(subscribersCache, producerService));
+        (RequestContext ctx_, Map<String, ?> data) -> {
+          AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+          if (ctx == null) {
+            return NoopFlow.INSTANCE;
+          }
+          while (true) {
+            DataSubscriberInfo subInfo = graphqlServerRequestMsgSubInfo;
+            if (subInfo == null) {
+              subInfo =
+                  producerService.getDataSubscribers(KnownAddresses.GRAPHQL_SERVER_ALL_RESOLVERS);
+              graphqlServerRequestMsgSubInfo = subInfo;
+            }
+            if (subInfo == null || subInfo.isEmpty()) {
+              return NoopFlow.INSTANCE;
+            }
+            DataBundle bundle =
+                new SingletonDataBundle<>(KnownAddresses.GRAPHQL_SERVER_ALL_RESOLVERS, data);
+            try {
+              GatewayContext gwCtx = new GatewayContext(true);
+              return producerService.publishDataEvent(subInfo, ctx, bundle, gwCtx);
+            } catch (ExpiredSubscriberInfoException e) {
+              graphqlServerRequestMsgSubInfo = null;
+            }
+          }
+        });
 
     subscriptionService.registerCallback(
-        EVENTS.databaseConnection(), new DatabaseConnectionCallback());
+        EVENTS.databaseConnection(),
+        (ctx_, dbType) -> {
+          AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+          if (ctx == null) {
+            return;
+          }
+          ctx.setDbType(dbType);
+        });
 
     subscriptionService.registerCallback(
-        EVENTS.databaseSqlQuery(), new DatabaseSqlQueryCallback(subscribersCache, producerService));
+        EVENTS.databaseSqlQuery(),
+        (ctx_, sql) -> {
+          AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+          if (ctx == null) {
+            return NoopFlow.INSTANCE;
+          }
+          while (true) {
+            DataSubscriberInfo subInfo = dbSqlQuerySubInfo;
+            if (subInfo == null) {
+              subInfo =
+                  producerService.getDataSubscribers(
+                      KnownAddresses.DB_TYPE, KnownAddresses.DB_SQL_QUERY);
+              dbSqlQuerySubInfo = subInfo;
+            }
+            if (subInfo == null || subInfo.isEmpty()) {
+              return NoopFlow.INSTANCE;
+            }
+            DataBundle bundle =
+                new MapDataBundle.Builder(CAPACITY_0_2)
+                    .add(KnownAddresses.DB_TYPE, ctx.getDbType())
+                    .add(KnownAddresses.DB_SQL_QUERY, sql)
+                    .build();
+            try {
+              GatewayContext gwCtx = new GatewayContext(false, RuleType.SQL_INJECTION);
+              return producerService.publishDataEvent(subInfo, ctx, bundle, gwCtx);
+            } catch (ExpiredSubscriberInfoException e) {
+              dbSqlQuerySubInfo = null;
+            }
+          }
+        });
   }
 
   public void stop() {
     subscriptionService.reset();
+  }
+
+  private static boolean hasUserTrackingEvent(final TraceSegment traceSeg) {
+    for (String tagName : USER_TRACKING_TAGS) {
+      final Object value = traceSeg.getTagTop(tagName);
+      if (value != null && "true".equalsIgnoreCase(value.toString())) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  private static void writeRequestHeaders(
+      final TraceSegment traceSeg,
+      final Set<String> allowed,
+      final Map<String, List<String>> headers) {
+    writeHeaders(traceSeg, "http.request.headers.", allowed, headers);
+  }
+
+  private static void writeResponseHeaders(
+      final TraceSegment traceSeg,
+      final Set<String> allowed,
+      final Map<String, List<String>> headers) {
+    writeHeaders(traceSeg, "http.response.headers.", allowed, headers);
+  }
+
+  private static void writeHeaders(
+      final TraceSegment traceSeg,
+      final String prefix,
+      final Set<String> allowed,
+      final Map<String, List<String>> headers) {
+    if (headers != null) {
+      headers.forEach(
+          (name, value) -> {
+            if (allowed.contains(name)) {
+              String v = String.join(",", value);
+              if (!v.isEmpty()) {
+                traceSeg.setTagTop(prefix + name, v);
+              }
+            }
+          });
+    }
+  }
+
+  private static class RequestContextSupplier implements Flow<AppSecRequestContext> {
+    private static final Flow<AppSecRequestContext> EMPTY = new RequestContextSupplier(null);
+
+    private final AppSecRequestContext appSecRequestContext;
+
+    public RequestContextSupplier() {
+      this(new AppSecRequestContext());
+    }
+
+    public RequestContextSupplier(AppSecRequestContext ctx) {
+      appSecRequestContext = ctx;
+    }
+
+    @Override
+    public Action getAction() {
+      return Action.Noop.INSTANCE;
+    }
+
+    @Override
+    public AppSecRequestContext getResult() {
+      return appSecRequestContext;
+    }
+  }
+
+  private static class NewRequestHeaderCallback
+      implements TriConsumer<RequestContext, String, String> {
+    @Override
+    public void accept(RequestContext ctx_, String name, String value) {
+      AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+      if (ctx == null) {
+        return;
+      }
+
+      if (name.equalsIgnoreCase("cookie")) {
+        Map<String, List<String>> cookies = CookieCutter.parseCookieHeader(value);
+        ctx.addCookies(cookies);
+      } else {
+        ctx.addRequestHeader(name, value);
+      }
+    }
+  }
+
+  private class RequestHeadersDoneCallback implements Function<RequestContext, Flow<Void>> {
+    public Flow<Void> apply(RequestContext ctx_) {
+      AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+      if (ctx == null || ctx.isReqDataPublished()) {
+        return NoopFlow.INSTANCE;
+      }
+      ctx.finishRequestHeaders();
+      return maybePublishRequestData(ctx);
+    }
+  }
+
+  private class MethodAndRawURICallback
+      implements TriFunction<RequestContext, String, URIDataAdapter, Flow<Void>> {
+    @Override
+    public Flow<Void> apply(RequestContext ctx_, String method, URIDataAdapter uri) {
+      AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
+      if (ctx == null) {
+        return NoopFlow.INSTANCE;
+      }
+
+      if (ctx.isReqDataPublished()) {
+        log.debug(
+            "Request method and URI already published; will ignore new values {}, {}", method, uri);
+        return NoopFlow.INSTANCE;
+      }
+      ctx.setMethod(method);
+      ctx.setScheme(uri.scheme());
+      if (uri.supportsRaw()) {
+        ctx.setRawURI(uri.raw());
+      } else {
+        try {
+          URI encodedUri = new URI(null, null, uri.path(), uri.query(), null);
+          String q = encodedUri.getRawQuery();
+          StringBuilder encoded = new StringBuilder();
+          encoded.append(encodedUri.getRawPath());
+          if (null != q && !q.isEmpty()) {
+            encoded.append('?').append(q);
+          }
+          ctx.setRawURI(encoded.toString());
+        } catch (URISyntaxException e) {
+          log.debug("Failed to encode URI '{}{}'", uri.path(), uri.query());
+        }
+      }
+      return maybePublishRequestData(ctx);
+    }
+  }
+
+  private Flow<Void> maybePublishRequestData(AppSecRequestContext ctx) {
+    String savedRawURI = ctx.getSavedRawURI();
+
+    if (savedRawURI == null || !ctx.isFinishedRequestHeaders() || ctx.getPeerAddress() == null) {
+      return NoopFlow.INSTANCE;
+    }
+
+    Map<String, List<String>> queryParams = EMPTY_QUERY_PARAMS;
+    int i = savedRawURI.indexOf("?");
+    if (i != -1) {
+      String qs = savedRawURI.substring(i + 1);
+      // ideally we'd have the query string as parsed by the server
+      // or at the very least the encoding used by the server
+      queryParams = parseQueryStringParams(qs, StandardCharsets.UTF_8);
+    }
+
+    String scheme = ctx.getScheme();
+    if (scheme == null) {
+      scheme = "http";
+    }
+
+    ctx.setReqDataPublished(true);
+
+    MapDataBundle bundle =
+        new MapDataBundle.Builder(CAPACITY_6_10)
+            .add(KnownAddresses.HEADERS_NO_COOKIES, ctx.getRequestHeaders())
+            .add(KnownAddresses.REQUEST_COOKIES, ctx.getCookies())
+            .add(KnownAddresses.REQUEST_SCHEME, scheme)
+            .add(KnownAddresses.REQUEST_METHOD, ctx.getMethod())
+            .add(KnownAddresses.REQUEST_URI_RAW, savedRawURI)
+            .add(KnownAddresses.REQUEST_QUERY, queryParams)
+            .add(KnownAddresses.REQUEST_CLIENT_IP, ctx.getPeerAddress())
+            .add(KnownAddresses.REQUEST_CLIENT_PORT, ctx.getPeerPort())
+            .add(KnownAddresses.REQUEST_INFERRED_CLIENT_IP, ctx.getInferredClientIp())
+            .build();
+
+    while (true) {
+      DataSubscriberInfo subInfo = this.initialReqDataSubInfo;
+      if (subInfo == null) {
+        subInfo =
+            producerService.getDataSubscribers(
+                KnownAddresses.HEADERS_NO_COOKIES,
+                KnownAddresses.REQUEST_COOKIES,
+                KnownAddresses.REQUEST_SCHEME,
+                KnownAddresses.REQUEST_METHOD,
+                KnownAddresses.REQUEST_URI_RAW,
+                KnownAddresses.REQUEST_QUERY,
+                KnownAddresses.REQUEST_CLIENT_IP,
+                KnownAddresses.REQUEST_CLIENT_PORT,
+                KnownAddresses.REQUEST_INFERRED_CLIENT_IP);
+        initialReqDataSubInfo = subInfo;
+      }
+
+      try {
+        GatewayContext gwCtx = new GatewayContext(false);
+        return producerService.publishDataEvent(subInfo, ctx, bundle, gwCtx);
+      } catch (ExpiredSubscriberInfoException e) {
+        this.initialReqDataSubInfo = null;
+      }
+    }
+  }
+
+  private Flow<Void> maybePublishResponseData(AppSecRequestContext ctx) {
+
+    int status = ctx.getResponseStatus();
+
+    if (status == 0 || !ctx.isFinishedResponseHeaders()) {
+      return NoopFlow.INSTANCE;
+    }
+
+    ctx.setRespDataPublished(true);
+
+    MapDataBundle bundle =
+        MapDataBundle.of(
+            KnownAddresses.RESPONSE_STATUS, String.valueOf(ctx.getResponseStatus()),
+            KnownAddresses.RESPONSE_HEADERS_NO_COOKIES, ctx.getResponseHeaders());
+
+    while (true) {
+      DataSubscriberInfo subInfo = respDataSubInfo;
+      if (subInfo == null) {
+        subInfo =
+            producerService.getDataSubscribers(
+                KnownAddresses.RESPONSE_STATUS, KnownAddresses.RESPONSE_HEADERS_NO_COOKIES);
+        respDataSubInfo = subInfo;
+      }
+
+      try {
+        GatewayContext gwCtx = new GatewayContext(false);
+        return producerService.publishDataEvent(subInfo, ctx, bundle, gwCtx);
+      } catch (ExpiredSubscriberInfoException e) {
+        respDataSubInfo = null;
+      }
+    }
+  }
+
+  private void maybeExtractSchemas(AppSecRequestContext ctx) {
+    boolean extractSchema = false;
+    if (Config.get().isApiSecurityEnabled() && requestSampler != null) {
+      extractSchema = requestSampler.sampleRequest();
+    }
+
+    if (!extractSchema) {
+      return;
+    }
+
+    while (true) {
+      DataSubscriberInfo subInfo = requestEndSubInfo;
+      if (subInfo == null) {
+        subInfo = producerService.getDataSubscribers(KnownAddresses.WAF_CONTEXT_PROCESSOR);
+        requestEndSubInfo = subInfo;
+      }
+      if (subInfo == null || subInfo.isEmpty()) {
+        return;
+      }
+
+      DataBundle bundle =
+          new SingletonDataBundle<>(
+              KnownAddresses.WAF_CONTEXT_PROCESSOR,
+              Collections.singletonMap("extract-schema", true));
+      try {
+        GatewayContext gwCtx = new GatewayContext(false);
+        producerService.publishDataEvent(subInfo, ctx, bundle, gwCtx);
+        return;
+      } catch (ExpiredSubscriberInfoException e) {
+        requestEndSubInfo = null;
+      }
+    }
+  }
+
+  private static Map<String, List<String>> parseQueryStringParams(
+      String queryString, Charset uriEncoding) {
+    if (queryString == null) {
+      return Collections.emptyMap();
+    }
+
+    Map<String, List<String>> result = new HashMap<>();
+
+    String[] keyValues = QUERY_PARAM_SPLITTER.split(queryString);
+
+    for (String keyValue : keyValues) {
+      String[] kv = QUERY_PARAM_VALUE_SPLITTER.split(keyValue, 2);
+      String value = kv.length > 1 ? urlDecode(kv[1], uriEncoding, true) : "";
+      String key = urlDecode(kv[0], uriEncoding, true);
+      List<String> strings = result.computeIfAbsent(key, k -> new ArrayList<>(1));
+      strings.add(value);
+    }
+
+    return result;
+  }
+
+  private static String urlDecode(String str, Charset charset, boolean queryString) {
+    return decodeString(str, charset, queryString, Integer.MAX_VALUE);
+  }
+
+  private static String decodeString(String str, Charset charset, boolean queryString, int limit) {
+    byte[] bytes = str.getBytes(charset);
+    int j = 0;
+    for (int i = 0; i < bytes.length && j < limit; i++, j++) {
+      int b = bytes[i];
+      if (b == 0x25 /* % */) {
+        if (i + 2 < bytes.length) {
+          int val = byteToDigit(bytes[i + 2]);
+          if (val >= 0) {
+            val += 16 * byteToDigit(bytes[i + 1]);
+            if (val >= 0) {
+              i += 2;
+              bytes[j] = (byte) val;
+              continue;
+            }
+          }
+        }
+      } else if (b == 0x2b /* + */ && queryString) {
+        bytes[j] = ' ';
+        continue;
+      }
+      bytes[j] = (byte) b;
+    }
+
+    return new String(bytes, 0, j, charset);
+  }
+
+  private static int byteToDigit(byte b) {
+    if (b >= 0x30 /* 0 */ && b <= 0x39 /* 9 */) {
+      return b - 0x30;
+    }
+    if (b >= 0x41 /* A */ && b <= 0x46 /* F */) {
+      return 10 + (b - 0x41);
+    }
+    if (b >= 0x61 /* a */ && b <= 0x66 /* f */) {
+      return 10 + (b - 0x61);
+    }
+    return -1;
   }
 
   private static class IGAppSecEventDependencies {
