--- dd-java-agent/appsec/src/main/java/com/datadog/appsec/gateway/callbacks/NewRequestHeaderCallback.java
@@ -1,26 +1,0 @@
-package com.datadog.appsec.gateway.callbacks;
-
-import com.datadog.appsec.gateway.AppSecRequestContext;
-import com.datadog.appsec.gateway.CookieCutter;
-import datadog.trace.api.function.TriConsumer;
-import datadog.trace.api.gateway.RequestContext;
-import datadog.trace.api.gateway.RequestContextSlot;
-import java.util.List;
-import java.util.Map;
-
-public class NewRequestHeaderCallback implements TriConsumer<RequestContext, String, String> {
-
-  @Override
-  public void accept(RequestContext ctx_, String name, String value) {
-    AppSecRequestContext ctx = ctx_.getData(RequestContextSlot.APPSEC);
-    if (ctx == null) {
-      return;
-    }
-    if (name.equalsIgnoreCase("cookie")) {
-      Map<String, List<String>> cookies = CookieCutter.parseCookieHeader(value);
-      ctx.addCookies(cookies);
-    } else {
-      ctx.addRequestHeader(name, value);
-    }
-  }
-}
